Create HeatMaps.xlsx
This excel file holds the data for the heat maps. To generate these maps, all you need to do is highlight the table and go to Insert->3D Maps->Open 3D Maps and in the graph type go to heat map. From there, it should auto complete the lat/lon and you place the frequency into the value. I found a radius of 5% is a good visualisation of the data ( it can be found under layer options)